+++ Assets/Prefabs/Han/EnemyType_1.prefab
@@ -248,7 +248,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   enemyType: 0
   speed: 5
-  hp: 5
+  hp: 10
   atk: 1
   maxHP: 10
   attackFarWithPlayer: 1.3
