+++ Assets/Prefabs/Han/EnemyType_3.prefab
@@ -278,7 +278,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   enemyType: 2
   speed: 5
-  hp: 5
+  hp: 10
   atk: 1
   maxHP: 10
   attackFarWithPlayer: 1.3
