+++ Assets/Prefabs/Han/TimeLine/GameOverShowTimeLine.playable.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ff5c31ef955d4744085095106cb56e14
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
