+++ Assets/Prefabs/Han/TimeLine/GameOverShowTimeLine2.playable.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ea765a592195f224bacf96affe4f54f4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
