+++ Assets/Prefabs/Han/TimeLine/GameOverShowTimeLine3.playable.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 68e3ed4e8f7984145966aa39debeb4b1
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
