+++ Assets/Scripts/Kan/EnemyAnimationEventForwarder.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2ac43e81fa4776f4bb518de1038c0385
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
