+++ Assets/Scripts/Kan/HPBar.cs
@@ -58,6 +58,11 @@ public class HPBar : MonoBehaviour
     private void Update()
     {
         
+        
+    }
+
+    private void LateUpdate()
+    {
         hpFillImage.fillAmount = Mathf.Clamp(hp * 1 / maxHP, 0, 1);
 
         transform.forward = Camera.main.transform.forward;
